pyproject.toml
[tool.poetry]
name = "hmt-bot"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fake-useragent = "^1.1.1"
requests ={extras = ["socks"],version = "^2.28.2"}
bs4 = "^0.0.1"
apscheduler = "^3.10.1"
python-telegram-bot = {extras = ["socks, job-queue"], version="^20.2"}

[tool.poetry.group.dev.dependencies]
jupyterlab = "3.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
